Create a git repo for site root if it does not exist

+++ test.sh
@@ -26,8 +26,14 @@ wp db export $TMP/before.sql
 # Switching to a new branch
 echo "* Switching to 'upgrade' branch"
 cd $SITEROOT
-git checkout -f # clean any unsaved changes
-git checkout master # checkout the master branch
+if [ -d "$SITEROOT" ]; then
+	git init
+	git add .
+	git commit -m 'First commit'
+else
+	git checkout -f # clean any unsaved changes
+	git checkout master # checkout the master branch
+fi
 git checkout -b upgrade || echo 'Could not create upgrade branch'; exit 1; # create a new upgrade branch, and exit if failed
 
 # Add our mu-plugin to collect 'error_log's
